// Front-end/src/resources/enums/enums.tsx
export enum EnumSecciones {
    Principal = 'principal',
    Iluminacion = 'iluminacion',
    Energia = 'energia',
    Aulas = 'aulas'
}

export enum IconName {
    Home = 'home',
    Info = 'info',
    AccountCircle = 'account_circle',
    Dashboard = 'dashboard',
    Add = 'add',
    Edit = 'edit',
    Temperatura = 'Temperatura',
    Temperatura_Obj = 'Temperatura_Obj',
    Humedad = 'Humedad',
    Luz = 'Luz',
    Bateria = 'Bateria',
    Movimiento = 'Movimiento',
    CO2 = 'CO2',
    Delete = 'Delete',
    Back = 'Back',
    Default = 'default',
    Null = 'null'
  }

export enum Tarjetas {
    Texto = 'Texto',
    Grafico = 'Grafico',
    Imagen = 'Imagen',
    Estado = 'Estado',
    Termostato = 'Termostato',
    Plano = 'Plano',
    Grupo = 'Grupo'
}

export enum TiempoGrafico {
    Dia = '24h',
    Semana = '7d',
    Mes = '30d',
    TMeses = '90d',
    Anio = '365d'
}

export enum TipoGrafico {
    Lineal = 'Lineal',
    Barra = 'Barra',
    Pastel = 'Pastel'
}

// export enum Posiciones {
//     CeroCero = '0x0',
//     CeroUno = '0x1',
//     CeroDos = '0x2',
//     UnoCero = '1x0',
//     UnoUno = '1x1',
//     UnoDos = '1x2',
//     DosCero = '2x0',
//     DosUno = '2x1',
//     DosDos = '2x2'
// }

// export enum Dimensiones {
//     UnoUno = '1x1',
//     UnoDos = '1x2',
//     UnoTres = '1x3',
//     DosUno = '2x1',
//     DosDos = '2x2',
//     DosTres = '2x3',
//     TresUno = '3x1',
//     TresDos = '3x2',
//     TresTres = '3x3'
// }

export enum Layouts{
    Card = 'c',
    Grid = 'g',
    Sidebar = 's'
}

export enum Comparadores {
    Mayor = '>',
    Menor = '<',
    Igual = '='
}

export enum Acciones {
    Set = 'SET'
}

export enum Planos {
    Fermin_Caballero_B = 'Planta_Baja',
    Fermin_Caballero_1 = 'Planta_1',
    Fermin_Caballero_2 = 'Planta_2',
    Anexo_B_1 = 'Anexo_B_1',
    Anexo_B_2 = 'Anexo_B_2',
}

export enum Espacios {
    Aula_A1_1 = 'Aula A1.1',
    Aula_A1_2 = 'Aula A1.2',
    Pasillo_A1 = 'Pasillo A1',
    Aula_A2_1 = 'Aula A2.1',
    Aula_A2_2 = 'Aula A2.2',
    Pasillo_A2 = 'Pasillo A2',
    LD1 = "LD1",
    LD2 = "LD2",
    LD3 = "LD3",
    LD4 = "LD4",
    Aula_F0_1 = "Aula F0.1",
    Aula_F1_1 = "Aula F1.1",
    Salon_De_Actos = "Salón de actos",
    Salon_De_Grados = "Salón de grados",
}